requirements.txt
playwright>=1.58.0
python-dotenv==1.0.1
httpx>=0.27.0
